+++ app/src/main/java/org/eyeseetea/malariacare/data/database/iomodules/local/importer/ImportController.java
@@ -3,22 +3,22 @@ package org.eyeseetea.malariacare.data.database.iomodules.local.importer;
 import android.content.Context;
 import android.net.Uri;
 
-import org.eyeseetea.malariacare.data.database.datasources.DBFlowLocalDataSource;
+import org.eyeseetea.malariacare.data.database.datasources.DatabaseImporter;
 import org.eyeseetea.malariacare.domain.boundary.IImportController;
 
 import java.io.IOException;
 
 public class ImportController implements IImportController {
-    DBFlowLocalDataSource dbFlowLocalDataSource;
+    DatabaseImporter mDatabaseImporter;
 
     public ImportController(Context context) {
-        dbFlowLocalDataSource = new DBFlowLocalDataSource(context);
+        mDatabaseImporter = new DatabaseImporter(context);
     }
 
     @Override
     public void importDB(Uri uri, IImportControllerCallback callback) {
         try{
-            dbFlowLocalDataSource.importDB(uri);
+            mDatabaseImporter.importDB(uri);
         }catch (IOException e){
             callback.onError(e);
             return;
